Function cargo_test_support::compare::match_exact

pub(crate) fn match_exact(
    expected: &str,
    actual: &str,
    description: &str,
    other_output: &str,
    cwd: Option<&Path>,
) -> Result<()>

Compares one string against another, checking that they both match.

See Patterns for more information on pattern matching.

description explains where the output is from (usually “stdout” or “stderr”).

other_output is other output to display in the error (usually stdout or stderr).