cargo_test_support::compare
Function match_with_without

pub(crate) fn match_with_without(
    actual: &str,
    with: &[String],
    without: &[String],
    redactions: &Redactions,
) -> Result<()>

Checks that the given string has a line that contains the given patterns, and that line also does not contain the without patterns.

See Patterns for more information on pattern matching.

See crate::Execs::with_stderr_line_without for an example and cautions against using.