cargo_test_support

Module containers

Support for testing using Docker containers.

The Container type is a builder for configuring a container to run. After you call launch, you can use the ContainerHandle to interact with the running container.

Tests using containers must use #[cargo_test(container_test)] to disable them unless the CARGO_CONTAINER_TESTS environment variable is set.

Structs

Container

A builder for configuring a container to run.

ContainerHandle

A handle to a running container.

MkFile

Builder for configuring a file to copy into a container.

Functions

remove_if_exists 🔒