Trait cargo_test_support::paths::CargoPathExt

pub trait CargoPathExt {
    // Required methods
    fn to_url(&self) -> Url;
    fn rm_rf(&self);
    fn mkdir_p(&self);
    fn ls_r(&self) -> Vec<PathBuf>;
    fn move_in_time<F>(&self, travel_amount: F)
       where F: Fn(i64, u32) -> (i64, u32);

    // Provided methods
    fn move_into_the_past(&self) { ... }
    fn move_into_the_future(&self) { ... }
}

Common path and file operations

Required Methods

fn to_url(&self) -> Url

fn rm_rf(&self)

fn mkdir_p(&self)

fn ls_r(&self) -> Vec<PathBuf>

Returns a list of all files and directories underneath the given directory, recursively, including the starting path.

fn move_in_time<F>(&self, travel_amount: F)
where F: Fn(i64, u32) -> (i64, u32),

Provided Methods

fn move_into_the_past(&self)

fn move_into_the_future(&self)

Object Safety

This trait is not object safe.