Module cargo_test_support::publish

Functions

_validate_upload ğŸ”’

create_index_line ğŸ”’

read_le_u32 ğŸ”’

split_index_features ğŸ”’

validate_alt_upload

Checks the result of a crate publish to an alternative registry.

validate_crate_contents

Checks the contents of a .crate file.

validate_upload

Checks the result of a crate publish.

validate_upload_with_contents

Checks the result of a crate publish, along with the contents of the files.

write_to_index ğŸ”’