Struct cargo_test_support::Project

pub struct Project {
    pub(crate) root: PathBuf,
}

A cargo project to run tests against.

See ProjectBuilder or Project::from_template to get started.

Fields
root: PathBuf
Implementations

impl Project

pub fn from_template(template_path: impl AsRef<Path>) -> Self

Copy the test project from a fixed state

pub fn root(&self) -> PathBuf

Root of the project, ex: /path/to/cargo/target/cit/t0/foo

pub fn build_dir(&self) -> PathBuf

Project’s target dir, ex: /path/to/cargo/target/cit/t0/foo/target

pub fn target_debug_dir(&self) -> PathBuf

Project’s debug dir, ex: /path/to/cargo/target/cit/t0/foo/target/debug

pub fn url(&self) -> Url

File url for root, ex: file:///path/to/cargo/target/cit/t0/foo

pub fn example_lib(&self, name: &str, kind: &str) -> PathBuf

Path to an example built as a library. kind should be one of: “lib”, “rlib”, “staticlib”, “dylib”, “proc-macro” ex: /path/to/cargo/target/cit/t0/foo/target/debug/examples/libex.rlib

pub fn bin(&self, b: &str) -> PathBuf

Path to a debug binary. ex: /path/to/cargo/target/cit/t0/foo/target/debug/foo

pub fn release_bin(&self, b: &str) -> PathBuf

Path to a release binary. ex: /path/to/cargo/target/cit/t0/foo/target/release/foo

pub fn target_bin(&self, target: &str, b: &str) -> PathBuf

Path to a debug binary for a specific target triple. ex: /path/to/cargo/target/cit/t0/foo/target/i686-apple-darwin/debug/foo

pub fn glob<P: AsRef<Path>>(&self, pattern: P) -> Paths

Returns an iterator of paths matching the glob pattern, which is relative to the project root.

pub fn change_file(&self, path: &str, body: &str)

Changes the contents of an existing file.

pub fn process<T: AsRef<OsStr>>(&self, program: T) -> Execs

Creates a ProcessBuilder to run a program in the project and wrap it in an Execs to assert on the execution. Example: p.process(&p.bin(“foo”)) .with_stdout(“bar\n”) .run();

pub fn cargo(&self, cmd: &str) -> Execs

Creates a ProcessBuilder to run cargo. Arguments can be separated by spaces. Example: p.cargo(“build –bin foo”).run();

pub fn rename_run(&self, src: &str, dst: &str) -> Execs

Safely run a process after cargo build.

Windows has a problem where a process cannot be reliably be replaced, removed, or renamed immediately after executing it. The action may fail (with errors like Access is denied), or it may succeed, but future attempts to use the same filename will fail with “Already Exists”.

If you have a test that needs to do cargo run multiple times, you should instead use cargo build and use this method to run the executable. Each time you call this, use a new name for dst. See rust-lang/cargo#5481.

pub fn read_lockfile(&self) -> String

Returns the contents of Cargo.lock.

pub fn read_file(&self, path: &str) -> String

Returns the contents of a path in the project root

pub fn uncomment_root_manifest(&self)

Modifies Cargo.toml to remove all commented lines.

pub fn symlink(&self, src: impl AsRef<Path>, dst: impl AsRef<Path>)