Module cargo_test_support::tools

Common executables that can be reused by various tests.

Statics

CLIPPY_DRIVER 🔒

ECHO 🔒

ECHO_WRAPPER 🔒

Functions

echo

Returns the path to an executable that prints its arguments.

echo_subcommand

Returns a project which builds a cargo-echo simple subcommand

echo_wrapper

Returns the path to an executable that works as a wrapper around rustc.

wrapped_clippy_driver

A wrapper around rustc instead of calling clippy.