cargo_test_support::registry
Struct RegistryBuilder

pub struct RegistryBuilder {
    alternative: Option<String>,
    token: Option<Token>,
    auth_required: bool,
    http_index: bool,
    http_api: bool,
    api: bool,
    configure_token: bool,
    configure_registry: bool,
    custom_responders: HashMap<String, Box<dyn Send + Fn(&Request, &HttpServer) -> Response>>,
    not_found_handler: Box<dyn Send + Fn(&Request, &HttpServer) -> Response>,
    delayed_index_update: usize,
    credential_provider: Option<String>,
}

Prepare a local TestRegistry fixture

See also init and alt_init

Fields
alternative: Option<String>

If set, configures an alternate registry with the given name.

token: Option<Token>

The authorization token for the registry.

auth_required: bool

If set, the registry requires authorization for all operations.

http_index: bool

If set, serves the index over http.

http_api: bool

If set, serves the API over http.

api: bool

If set, config.json includes ‘api’

configure_token: bool

Write the token in the configuration.

configure_registry: bool

Write the registry in configuration.

custom_responders: HashMap<String, Box<dyn Send + Fn(&Request, &HttpServer) -> Response>>

API responders.

not_found_handler: Box<dyn Send + Fn(&Request, &HttpServer) -> Response>

Handler for 404 responses.

delayed_index_update: usize

If nonzero, the git index update to be delayed by the given number of seconds.

credential_provider: Option<String>

Credential provider in configuration

Implementations

impl RegistryBuilder

pub fn new() -> RegistryBuilder

pub fn add_responder<R: 'static + Send + Fn(&Request, &HttpServer) -> Response>( self, url: impl Into<String>, responder: R, ) -> Self

Adds a custom HTTP response for a specific url

pub fn not_found_handler<R: 'static + Send + Fn(&Request, &HttpServer) -> Response>( self, responder: R, ) -> Self

pub fn delayed_index_update(self, delay: usize) -> Self

Configures the git index update to be delayed by the given number of seconds.

pub fn alternative_named(self, alt: &str) -> Self

Sets whether or not to initialize as an alternative registry.

pub fn alternative(self) -> Self

Sets whether or not to initialize as an alternative registry.

pub fn no_configure_token(self) -> Self

Prevents placing a token in the configuration

pub fn no_configure_registry(self) -> Self

Prevents adding the registry to the configuration.

pub fn token(self, token: Token) -> Self

Sets the token value

pub fn auth_required(self) -> Self

Sets this registry to require the authentication token for all operations.

pub fn http_index(self) -> Self

Operate the index over http

pub fn http_api(self) -> Self

Operate the api over http

pub fn no_api(self) -> Self

The registry has no api.

pub fn credential_provider(self, provider: &[&str]) -> Self

The credential provider to configure for this registry.

pub fn build(self) -> TestRegistry

Initializes the registry.