cargo_test_support
Trait ChannelChangerCommandExt

pub trait ChannelChangerCommandExt {
    // Required method
    fn masquerade_as_nightly_cargo(self, _reasons: &[&str]) -> Self;
}

Enable nightly features for testing

Required Methods

fn masquerade_as_nightly_cargo(self, _reasons: &[&str]) -> Self

The list of reasons should be why nightly cargo is needed. If it is because of an unstable feature put the name of the feature as the reason, e.g. &["print-im-a-teapot"].

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.