Trait cargo_test_support::ChannelChangerCommandExt

pub trait ChannelChangerCommandExt {
    // Required method
    fn masquerade_as_nightly_cargo(self, _reasons: &[&str]) -> Self;
}

Enable nightly features for testing

Required Methods

fn masquerade_as_nightly_cargo(self, _reasons: &[&str]) -> Self

The list of reasons should be why nightly cargo is needed. If it is because of an unstable feature put the name of the feature as the reason, e.g. &["print-im-a-teapot"].

Object Safety

This trait is not object safe.